+++ lithops/serverless/backends/k8s/entry_point.py
@@ -57,9 +57,9 @@ def get_range(jobkey, total_calls, chunksize):
 
 def run_master_server():
     # Start Redis Server in the background
-    logger.info("Starting redis server in Master Pod")
-    os.system("redis-server --bind 0.0.0.0 --daemonize yes")
-    logger.info("Redis server started")
+    # logger.info("Starting redis server in Master Pod")
+    # os.system("redis-server --bind 0.0.0.0 --daemonize yes")
+    # logger.info("Redis server started")
 
     proxy.logger.setLevel(logging.DEBUG)
     proxy.run(debug=True, host='0.0.0.0', port=config.MASTER_PORT, use_reloader=False)
@@ -104,7 +104,7 @@ def run_job(payload):
             try:
                 server = f'http://{master_ip}:{config.MASTER_PORT}'
                 url = f'{server}/get-range/{job_key}/{total_calls}/{chunksize}'
-                res = requests.get(url)
+                res = requests.get(url, timeout=0.1)
                 call_ids_range = res.text  # for example: 0-5
             except Exception:
                 time.sleep(0.1)